backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
supabase>=2.10.0
python-dotenv>=1.0.0
httpx>=0.27.0
PyJWT>=2.9.0
anthropic>=0.40.0
